+++ test_app/client_tests.py
@@ -40,7 +40,7 @@ def test_get_author(pact, library_client: LibraryClient):
     expected = Like({"id": 1, "name": "Blake Crouch"})
 
     (
-        pact.given("An Author with ID 1 exists")
+        pact.given("An Author", id=1)
         .upon_receiving("A request for author ID 1")
         .with_request("GET", "/authors/1")
         .will_respond_with(200, body=expected)
@@ -52,7 +52,7 @@ def test_get_author(pact, library_client: LibraryClient):
 
 def test_update_author(pact, library_client: LibraryClient):
     (
-        pact.given("An Author with ID 1 exists")
+        pact.given("An Author", id=1)
         .upon_receiving("An author update request")
         .with_request("PATCH", "/authors/1", body={"name": "Helene Wecker"})
         .will_respond_with(200, body={"name": "Helene Wecker", "id": Like(1)})
@@ -64,7 +64,7 @@ def test_update_author(pact, library_client: LibraryClient):
 
 def test_delete_author(pact, library_client: LibraryClient):
     (
-        pact.given("An Author with ID 1 exists")
+        pact.given("An Author", id=1)
         .upon_receiving("An author deletion request")
         .with_request("DELETE", "/authors/1")
         .will_respond_with(204)
@@ -89,7 +89,7 @@ def test_get_books_by_author_no_author(pact, library_client: LibraryClient):
 def test_get_books_by_author_success(pact, library_client: LibraryClient):
     expected = EachLike({"id": 1, "title": "Dune"})
     (
-        pact.given("An Author with ID 1 exists")
+        pact.given("An Author", id=1)
         .and_given("A Book exists with author ID 1")
         .upon_receiving("A book search request for author ID 1")
         .with_request("GET", "/books", query={"authorId": ["1"]})