+++ test_app/provider/tests/test_pact.py
@@ -5,8 +5,8 @@ from pact_testgen.public import Response, verify_response
 
 from .provider_states import (
     setup_nothing,
-    setup_an_author_with_id_1_exists,
-    setup_an_author_with_id_1_exists_a_book_exists_with_author_id_1,
+    setup_an_author_id_1,
+    setup_an_author_id_1_a_book_exists_with_author_id_1,
 )
 
 
@@ -62,9 +62,9 @@ class TestNothing(TestCase):
         result.assert_success()
 
 
-class TestAnAuthorWithId1Exists(TestCase):
+class TestAnAuthorId1(TestCase):
     def setUp(self):
-        setup_an_author_with_id_1_exists()
+        setup_an_author_id_1()
 
     def a_request_for_author_id_1(self):
         raw_actual_response = self.client.generic(
@@ -140,9 +140,9 @@ class TestAnAuthorWithId1Exists(TestCase):
         result.assert_success()
 
 
-class TestAnAuthorWithId1ExistsABookExistsWithAuthorId1(TestCase):
+class TestAnAuthorId1ABookExistsWithAuthorId1(TestCase):
     def setUp(self):
-        setup_an_author_with_id_1_exists_a_book_exists_with_author_id_1()
+        setup_an_author_id_1_a_book_exists_with_author_id_1()
 
     def a_book_search_request_for_author_id_1(self):
         raw_actual_response = self.client.generic(